Work around fake Git history on Github actions
When our test suite runs on Github actions, the Git history is fake. We know
this because Git on Github actions reports README.md to be created two years
after its inception when run on CI:

    https://github.com/mikrobloggeriet/mikrobloggeriet.no/actions/runs/20596871353/job/59153283701

Our workaround is not to run the offending test when we can't ensure that we're
in a safe environment, such as `:dev` or `:prod`.